PHPStan doesn't know about psalm `empty` (#104)
* PHPStan doesn't know about psalm `empty`

This error was slient until 0.12.20 so we have to
have a specific return for each tool as suggested
by maintainer: https://github.com/phpstan/phpstan/issues/3253

As the change is not compatible with 0.12.19, I choose
to bump phpstan in the same time. The purpose of this change 
is to bump anyway.