Do not render metrics with incoherent labels
This contribution tries to prevent situations where the rendered
information could be incorrect.
If a new label is added (or removed) to a metric and the datastore is
not flushed after the addition (or removal) it is not possible to
determine the correspondance between the label name and a value (at
least not with the current design of the data stores).

Instead of throwing warnings and trying to still render the information,
we should stop and give the opportunity to fix the inconsistency of the
data store.